+++ UI.ER.AvaloniaUI.Test/RegistreDITest.cs
@@ -103,6 +103,7 @@ namespace UI.ER.AvaloniaUI.Test
             Type[] necessaris =
             [
                 typeof(AppStatusViewModel), typeof(UtilitatsViewModel),
+                typeof(DadesUsuariViewModel),
                 typeof(ActuacioSetViewModel), typeof(AlumneSetViewModel),
                 typeof(CentreSetViewModel), typeof(CursAcademicSetViewModel),
                 typeof(EtapaSetViewModel), typeof(TipusActuacioSetViewModel),